+++ cuda/selections/Hlt1/src/TrackMVALines.cu
@@ -9,8 +9,8 @@ namespace TrackMVALines {
     decision &=
       ((ptShift > maxPt && track.ipChi2 > minIPChi2) ||
        (ptShift > minPt && ptShift < maxPt &&
-        std::log(track.ipChi2) >
-          param1 / (ptShift - param2) / (ptShift - param2) + param3 / maxPt * (maxPt - ptShift) + std::log(minIPChi2)));
+        logf(track.ipChi2) >
+          param1 / (ptShift - param2) / (ptShift - param2) + param3 / maxPt * (maxPt - ptShift) + logf(minIPChi2)));
     return decision;
   }
